MCMC-Smapling
Implemented crude MCMC loop for BDE-fitting.

+++ bde/ml/datasets.py
@@ -263,8 +263,8 @@ class DatasetWrapper(BasicDataset):
         # TODO: Complete
         """
         chex.assert_equal(x.shape[0], y.shape[0])
-        self.x = x
-        self.y = y
+        self.x = x.astype(float)  # TODO: Needs better handling
+        self.y = y.astype(float)
         self._batch_size = batch_size
         self._seed = seed
